Simplify layout selection
We note that we mostly call `merge_all` with a single item list, sometimes with
2 items. So, we specialize.

Further, we assume that all sublists are sorted. If they are, then we can apply
the given simplifications.

Note that this does give up on one interesting property which was removal of
duplicates. It would be interesting to try and recover that.